chunk size 整型调整
CreateUpload时只检查文件md5是否存在&对已存在的文件上传分片时若分片已存在序号已存在则返回下一需要的分片号

add print

print http serialize time

去掉刚才加的打印时间的代码

UploadChunk调整为纯二进制流模式

readme

+++ internal/biz/uploading.go
@@ -5,19 +5,19 @@ import "github.com/google/uuid"
 type Uploading struct {
 	Upid       string `json:"upid"`
 	Filename   string `json:"filename"`
-	TotalSize  int32  `json:"total_size"`
+	TotalSize  int64  `json:"total_size"`
 	MD5SUM     string `json:"md5sum"`
 	Path       string `json:"path"`
-	CurrentNum int32  `json:"current_num"`
-	TotalCount int32  `json:"total_count"`
+	CurrentNum int64  `json:"current_num"`
+	TotalCount int64  `json:"total_count"`
 }
 
 // Chunk is a Chunk model.
 type Chunk struct {
 	CID  int    `json:"cid"`
 	Upid string `json:"upid"`
-	Num  int32  `json:"num"`
-	Size int32  `json:"size"`
+	Num  int64  `json:"num"`
+	Size int64  `json:"size"`
 	Path string `json:"path"`
 }
